Split matches.py to entire directory of match type definitions
