+++ server/collab/matches/assembly_hash.py
@@ -1,0 +1,6 @@
+import hash_match
+
+
+class AssemblyHashMatch(hash_match.HashMatch):
+  vector_type = 'assembly_hash'
+  match_type = 'assembly_hash'
