+++ server/collab/matches/match.py
@@ -1,0 +1,5 @@
+class Match:
+  @classmethod
+  def match(cls, source, target):
+    raise NotImplementedError("Method match for vector type {} not "
+                              "implemented".format(cls))
