+++ server/collab/matches/mnemonic_hash.py
@@ -1,0 +1,6 @@
+import hash_match
+
+
+class MnemonicHashMatch(hash_match.HashMatch):
+  vector_type = 'mnemonic_hash'
+  match_type = 'mnemonic_hash'
