+++ server/collab/matches/mnemonic_hist.py
@@ -1,0 +1,6 @@
+import hist_match
+
+
+class MnemonicHistogramMatch(hist_match.HistogramMatch):
+  vector_type = 'mnemonic_hist'
+  match_type = 'mnemonic_hist'
