+++ server/collab/matches/opcode_hist.py
@@ -1,0 +1,6 @@
+import hist_match
+
+
+class OpcodeHistogramMatch(hist_match.HistogramMatch):
+  vector_type = 'opcode_histogram'
+  match_type = 'opcode_histogram'
